Add annotation property to ChatStreamResult.Choice.ChoiceDelta
